requirements.txt
python-dotenv
streamlit
opencv-python
google-generativeai
faiss-cpu
scikit-image
matplotlib
toml
openpyxl
Pillow 